cloud functions/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
Pillow
pillow-heif